+++ NEWS.md
@@ -1,5 +1,6 @@
 # lasR 0.21.0
 
+- New: `local_maximum()` and `local_maximum_raster()` now accept `ws` as a function for variable-size moving windows.
 - New: Support of Entwine Point Tile format
 - New: `classify_with_ivf` gains a parameter `filter` (#289)
 - Change: `drop_noise()` and `keep_noise()` now filter classes 7 and 18 instead of 18 only (#283)